Move  into utils package
