+++ pytorch_grad_cam/gradcam.py
@@ -2,7 +2,6 @@ import cv2
 import numpy as np
 import torch
 from torch.autograd import Function
-from torchvision import transforms
 
 
 class FeatureExtractor:
@@ -54,16 +53,6 @@ class ModelOutputs:
                 x = module(x)
 
         return target_activations, x
-
-
-def preprocess_image(img):
-    normalize = transforms.Normalize(mean=[0.485, 0.456, 0.406],
-                                     std=[0.229, 0.224, 0.225])
-    preprocessing = transforms.Compose([
-        transforms.ToTensor(),
-        normalize,
-    ])
-    return preprocessing(img.copy()).unsqueeze(0)
 
 
 class GradCam:
