+++ pytorch_grad_cam/utils/image.py
@@ -1,5 +1,16 @@
 import cv2
 import numpy as np
+from torchvision import transforms
+
+
+def preprocess_image(img):
+    normalize = transforms.Normalize(mean=[0.485, 0.456, 0.406],
+                                     std=[0.229, 0.224, 0.225])
+    preprocessing = transforms.Compose([
+        transforms.ToTensor(),
+        normalize,
+    ])
+    return preprocessing(img.copy()).unsqueeze(0)
 
 
 def deprocess_image(img):
